requirements.txt
Django==4.2.20
djangorestframework==3.15.2
django-filter==25.1
djoser==2.3.1
drf-extra-fields==3.7.0
